django-experience/pyproject.toml
[project]
name = "django-experience"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "django-ninja>=1.3.0",
    "django>=4.2.16",
    "djangorestframework>=3.15.2",
]
